+++ packages/common/src/contractHelpers.ts
@@ -1,4 +1,3 @@
-import { Buffer } from 'buffer/';
 import { ContractAddress, InstanceInfo } from './types';
 
 const CONTRACT_PARAM_MAX_LENGTH = 65535;
@@ -25,8 +24,8 @@ export const getContractName = ({ name }: InstanceInfo): string => {
  *
  * @throws If buffer exceeds max length allowed for smart contract parameters
  */
-export const checkParameterLength = (buffer: Buffer): void => {
-    if (buffer.length > CONTRACT_PARAM_MAX_LENGTH) {
+export const checkParameterLength = (buffer: ArrayBuffer): void => {
+    if (buffer.byteLength > CONTRACT_PARAM_MAX_LENGTH) {
         throw new Error(
             `Serialized parameter exceeds max length of smart contract parameter (${CONTRACT_PARAM_MAX_LENGTH} bytes)`
         );
@@ -43,15 +42,11 @@ export const isEqualContractAddress =
 
 /** The name of a smart contract. Note: This does _not_ including the 'init_' prefix. */
 export type ContractName = string;
-/** The name of a receive function exposed in a smart contract module. Note: This is of the form '<contractName>.<entrypointName>'. */
-export type ReceiveName = string;
-/** The name of an init function exposed in a smart contract module. Note: This is of the form 'init_<contractName>'. */
-export type InitName = string;
 /** The name of an entrypoint exposed by a smart contract. Note: This does _not_ include the '<contractName>.' prefix. */
 export type EntrypointName = string;
 
 /** Check that every character is an Ascii alpha, numeric or punctuation. */
-function isAsciiAlphaNumericPunctuation(string: string) {
+export function isAsciiAlphaNumericPunctuation(string: string): boolean {
     for (let i = 0; i < string.length; i++) {
         const charCode = string.charCodeAt(i);
         if (
@@ -72,7 +67,7 @@ function isAsciiAlphaNumericPunctuation(string: string) {
 }
 
 /** Check if a string is a valid smart contract init name. */
-export function isInitName(string: string): string is InitName {
+export function isInitName(string: string): boolean {
     return (
         string.length <= 100 &&
         string.startsWith('init_') &&
@@ -82,12 +77,12 @@ export function isInitName(string: string): string is InitName {
 }
 
 /** Get the contract name from a string. Assumes the string is a valid init name. */
-export function getContractNameFromInit(initName: InitName): ContractName {
+export function getContractNameFromInit(initName: string): ContractName {
     return initName.substring(5);
 }
 
 /** Check if a string is a valid smart contract receive name. */
-export function isReceiveName(string: string): string is ReceiveName {
+export function isReceiveName(string: string): boolean {
     return (
         string.length <= 100 &&
         string.includes('.') &&
@@ -96,7 +91,7 @@ export function isReceiveName(string: string): string is ReceiveName {
 }
 
 /** Get the contract name and entrypoint name from a string. Assumes the string is a valid receive name. */
-export function getNamesFromReceive(receiveName: ReceiveName): {
+export function getNamesFromReceive(receiveName: string): {
     contractName: ContractName;
     entrypointName: EntrypointName;
 } {